Fix possible overflow for large sizes
To compute the length of the output vector, the product of the input dimensions was used. However, the product used i32, while the length of the Vec is usize. Hence, the code artificially restricted inputs to have a total of 2^31 elements, as opposed to 2^31 elements per dimension (as imposed by hptt).